Fix casting

+++ Server/Web/Structures/Settings/ServerDebugSettings.cs
@@ -4,8 +4,8 @@ namespace Server.Web.Structures.Settings
 {
     public class ServerDebugSettings
     {
-        public int SimulatedLossChance => DebugSettings.SettingsStore.SimulatedLossChance;
-        public int SimulatedDuplicatesChance => DebugSettings.SettingsStore.SimulatedDuplicatesChance;
+        public float SimulatedLossChance => DebugSettings.SettingsStore.SimulatedLossChance;
+        public float SimulatedDuplicatesChance => DebugSettings.SettingsStore.SimulatedDuplicatesChance;
         public int MaxSimulatedRandomLatencyMs => DebugSettings.SettingsStore.MaxSimulatedRandomLatencyMs;
         public int MinSimulatedLatencyMs => DebugSettings.SettingsStore.MinSimulatedLatencyMs;
     }
